moves ops to better folder
