+++ galaxy2galaxy/layers/spectral_ops.py
